migrating reports, raw data, processed data and py files
