--- brat-v1.3_Crunchy_Frog/data/first-iter/balanced_dataset_brat.txt
@@ -1,84 +1,0 @@
-#sinluz  terrazas  del  club  hipico      calle  bolivia  
-
-#sinluz  valle  de  la  pascua  desde  las  7  00  am
-
-20  horas        #sinluz
-
-hoy  toca  ramen  a  la  luz  de  una  vela  #sinluz  #lptm
-
-dios  si  nos  devolves  la  loooz      me  vuelvo  provida    heterosexual      catolico  apostolico  romano  y  practicamente      homofobico  y  machirulo      te  va  todo  eso      #sinluz
-
-que  gran  invento  el  celular  #sinluz
-
-#sinluz  esta  usted  y  su  falla  debe  ser  neurologica  ahora  apoya  a  cfk      pregunta  como  anda  su  pyme  de  35  asesores        
-
-4  veces  con  #apagon  en  varios  sectores  de    #laasuncion  #sinluz    7  15pm  #5julio  @rconfidencial  @bajolalupafm
-
-#sinluz  los  caobos    sabana  grande    la  florida    la  campigna        se  fue  por  fases  porque  hay  edificios  que  si  tienen  energia  electrica
-
-#22nov  #guayana  alta  vista    los  olivos    castillito    villa  colombia    villa  asia    villa  alianza    villa  brasil  #sinluz
-
-@vpitv  @mariacquintero  san  francisco  #sinluz
-
-15  horas  #sinluz  y  de  paso  sin  gas    y  si  eso  no  fuera  suficiente  3  meses  y  19  dias  #sinagua  agoniza  la  mal  querida  paraguana    #20s
-
-@luisgonzaloprz  #sinluz  desde  el  estado  lara
-
-@lfp1875  @hectoryb  ya  son  4  horas  #sinluz  en  #ciudadbolivar  y  semanas  #sinagua  10  53  pm
-
-#singasolina@dmurolo  la  cola  llega  al  km18  #pnm  imagino  sera  por  la  bomba  gasolina
-
-en  walker  martinez  con  tobalaba    la  florida  #sinluz
-
-colinas  de  bello  monte  #sinluz  12  25  pm
-
-madrugada  #sinluz  y  magnana  a  despertarse  para  hacer  la  cola  de  la  #gasolina  no  habra  un  almuerzo  decente  al  retornar  con  suerte  alrededor  de  las  3  o  4pm  y  el  tanque  a  medio  llenar  con  solo  30  litros    puede  que  si  ya  llego  la  luz    logre  cenar  algo  caliente
-
-@traffivalencia  en  que  bomba  hay  gasolina  #valencia  #singasolina    #16dic
-
-pzo  #singasolinamcb  #singasolina  cbv  #singasolina  ptl  #singasolina#ccs  #congasolina    que  cosas  no
-
-@vtvcanal8  @emmacaroagurto  @michelcaballero  en  san  cristobal  estamos  #sinluz  desde  las  6am    y  los  apagones  desde  marzo    #singas  desde  enero    magnana  estaremos  #sinagua    ya  lo  anunciaron  y  #singasolina  de  manera  normal    desde  hace  agnos
-
-#ahora  estaciones  esperando  por  la  cisterna  en  pto  ordaz    morochas  av    atlantico    la  de  los  olivos  villa  africana    las  americas  subiendo  a  alta  vista  y  la  que  esta  frente  a  c  alta  vista  2    #singasolina  #guayana  10  28  am
-
-a  las  5  44  am    aun  no  ha  llegado  la  gandola  con  gasolina    no  sabemos  a  que  hora  comenzaran  a  trabajar  los  empleados  de  la  estacion  de  servicio    asi  se  vive  #singasolina
-
-#28dic  #singasolina    #ciudadguayana  deberian  ver  las  colas  aqui  haciendo  una  desde  las  6  de  la  tarde  el  viernes  27  para  equipar  el  lunes  30  ya  hay  aproximadamente  300  carros  y  sigu    
-
-valera  estado  trujillo  ya  cuenta  con  servicio  electrico    #sinluz  falta  el  agua    el  gas    la  gasolina    la  seguridad  la  salud    educacion    transporte    pero  por  algo  se  empieza      
-
-no  soporto  el  caos  que  se  ha  vuelto  mi  maracaibo    el  calor  esta  brutal  y  #sinluz    no  puedo  moverme  porque  no  hay  gasolina    tengo  muy  poca  y  la  unica  estacion  que  vi  abierta  las  colas  kilometricas    de  verdad  hoy  no  tengo  paciencia
-
-#1abr  en  foto  la  fuerte  cola  que  se  registro  en  horas  de  la  magnana  en  la  bomba  de  gasolina  en  la  parroquia  #lavega  sector  #montalban  #caracas    en  el  distribuidor  la  yaguara  #singasolina    
-
-hoy  me  desperte  y  amanecimos  #sinagua  #singasolina  #singas  yo  me  pregunto  en  una  sociedad  tan  moderna  en  venezuela  pasamos  estas  penurias        #queremosvidadigna  #mujeresdeltachira
-
-mas  del  75%  de  maracaibo  se  encuentra  a  esta  hora  #sinluz    a  la  penumbra  se  suman  varios  municipios  del  zulia    si  usted  esta  a  esta  hora  sin  luz    no  tiene  agua    gas  ni  gasolina  para  moverse    organice  una  protesta  en  su  comunidad  y  que  el  pueblo  arrecho  reclame  sus  derechos
-
-#denuncia  en  varias  parroquias  del  oeste  de  #caracas  se  reportan  largas  colas  para  surtir  de  gasolina  y  hay  otras  donde  se  encuentran  cerrada  por  estar  #singasolina    la  crisis  del  combustible  llego  a  la  capital    por  culpa  de  un  regimen  corrupto  increible  y  somos  un  pais  petrolero
-
-#sinluzencuarentena  #sinagua  #singasolina  asi  tenemos  que  aguantar  el  #coronavirusvenezuela  #covid19
-
-#carabobo  #singasolina  imposible  salir  en  vehiculos  a  buscar  alimentos  y  agua  potable    #venezuela  tierra  de  zombies  #thewalkingdead
-
-en  valle  de  la  pascua  la  situacion  de  la  gasolina  es  alarmante    #singasolina
-
-#sinluz  esta  usted  y  su  falla  debe  ser  neurologica  ahora  apoya  a  cfk      pregunta  como  anda  su  pyme  de  35  asesores        
-
-#sinluz  #lacabagna  #laboyera  #elhatillo  @elhatillo  @elhatillano  @eliasayegh  desde    hace  una  hora
-
-buenas  tardes  @corpoelecinfo  se  mantiene  el  no  suministro  de  energia  electrica  en  la  esquina  de  delicias  parroquia  san  juan  caracas    seguimos  #sinluz  quedamos  atentos  a  sus  gestiones  gracias
-
-asi  desalojan  el  centro  comercial  san  ignacio  por  el  #apagon      este  se  trata  de  la  peor  falla  electrica  desde  el  pasado  10  de  abril  #sinluz  video    cortesia    
-
-@corpoeleccar_  7  hrs  continuas  #sinluz  y  nada  
-
-@willy_casanova  @nicolasmaduro  que  paso  con  hidrolago  no  bombea  agua  en  cuatricentenario  desde  hace  12  dias  los  cortes  #sinluz    estan  a  la  orden  del  dia  necesitamos  agua  por  favor
-
-@corpoeleczulia_  @fbritomaestre  @drodriguezven  #sinluz  #sinluz  @corpoeleczulia_    parroquia  chiquinquira  sin  electricidad    sector  grano  de  oro  #maracaibo  #sinluz  #corpoeleczulia_  desde  las  07  10  pm  #nomasracionamiento
-
-y  ahora  y  para  variar  desde  las  2pm  #sinluz  #tachira  #sancristobal
-
-@corpoeleczulia_  dios  mioo  que  pasa  sigue  el  viacrusis  desde  840pm  #sinluz  en  circuito  las  marias  pongan  la  luz  por  favor  el  calor  es  insoportable  @corpoelecinfo  @fbritomaestre  @mppeevzla  @omarprietogob  @diariopanorama
-
