--- brat-v1.3_Crunchy_Frog/data/first-iter/sampled_58_30.txt
@@ -1,60 +1,0 @@
-y  continuamos  "disfrutando"  de  los  regalos  navidegnos  de  @corpoelecinfo  y  @corpoeleclara  mas  raciones  de  navidades  #sinluz  hoy  #25dic  de  verdad  que  la  supieron  hacer    felicitaciones!  nunca  nos  defraudan      
-
-ya  van  casi  tres  horas  #sinluz  en  el  rectorado  uc    la  lluvia  paro    ya  el  sol  seco  las  calles    pero  la  incapacidad  socialista  sigue      #socialismoesmiseria  #chavismoescrimen  #madurismoesestupidez    #lacavaescomplice  #menosmalqueyasevan
-
-tops  deportivos  disponibles  hermosos  super  comodos      envios  nacionales        #nuevamercancia  #coleccion  2020  disegnos  unicos  %    ropa  colombiana  precios  al  whatsapp  bio  0412    3185082#licras  #moda2020  #sinluz  #tops  #topsfashion  #sportwoman  #gym  #womanwear  #caracas    
-
-aunque  parezca  dificil  en  momentos  como  estos  es  cuando  mas  necesitamos  de  ella    no  dejes  que  nada  apague  tu  sonrisa  aunque  estes  #sinluz    animo!estamos  contigo  #sonrisascagnadeazucar  #maracaysinluz  #venezuelasinluz  #animo    
-
-desde  la  magnana  problemas  de  tension  en  #almagro  #sinluz
-
-actualizado    usuarios  sin  suministro  electrico    @oficialedesur  @edenorclientes  #sinluz  #secorto    
-
-#singasolina  |  cola  para  surtir  de  combustible  en  san  cristobal    estado  tachira    este  estado  sufre  de  racionamiento  de  gasolina  desde  hace  agnos    exponiendo  a  sus  ciudadanos  a  permanecer  largas  horas    a  veces  dias  para  poder  abastecer  el  vehiculo    video  @mariananduque    
-
-#2ene  #portuguesa  #singasolina  manuel  eduardo  silva      conductor    "llegamos  despues  del  feliz  agno    nos  abocamos  a  hacer  la  cola  para  surtir  gasolina  porque  necesitamos  transportarnos    soy  comerciante"        @vpitv      
-
-#sinluz  en  #guarenas  nueva  casarapa    camino  real    30  horas  sin  luz    esto  es  inhumano  
-
-@eliaspino  @chuotorrealba  no  puedo  verlo  estoy  en  #maracaibo  #sinluz  esperando  el  cese  de  la  usurpacion  y  bla  bla  bla  bla  las  otras  guevonadas  !!!
-
-increible  que  #23dic  y  la  dictadura  ni  una  palabra  para  explicar  que  #venezuela  esta  #singasolina  y  los  racionamientos  electricos  son  mas  extenso    navidad  feliz    sin  duda  que  no  
-
-[apocalipsis  rojo]  10  fotos  que  evidencian  el  drama  que  vivio  nuevamente  una  venezuela  #sinluz    
-
-alguien  vio  la  pelicula  "en  lo  profundo  del  bosque"    esta  pasando    #apocalipsis  #sinluz
-
-a  ver        ayer  #14ago  en  #barinas  hubieron  4  cortes  electricos    hoy  #15ago    dio  inicio  el  dia  #sinluz    entonces      #venezuelaavanzayvence  huele  a  bots  circo  comunistas  #sosvenezuela  nos  estan  matando    no  hagan  la  vista  gorda!      
-
-que  co#o  esperan  los  venezolanos  para  formar  el  verguero  en  las  calles!  #singasolina    #singas  #sinluz  sueldo  minimo  5  dolare$  mensuales    hospitales  y  transporte  publico  destartalado      pero  la  culpa  es  de  @jguaido      ya  estamos  amaestrados!  como  cuba!  cansa    echale  pichon!    
-
-longaray  el  valle  en  caracas  #sinluz  desde  las  5  30  aproximadamente    s  bueeeeenoooo  ojala  en  vez  de  radares  militares  hubiesen  puesto  cables  y  transformadores  ps  
-
-en  san  carlos    cojedes  se  restauro  el  servicio  electrico    no  se  como  estaran  los  demas  despues  del  gran  bajon  #sinluz  #apagon
-
-#21jul  |  apagones  dejan  sin  agua  a  municipios  de  trujillo  &gt;&gt;    
-
-#19jul  |  falla  en  subestacion  yaracuy  afecto  suministro  electrico  en  maracaibo  &gt;&gt;    
-
-20  puntos  para  rafael  orozco!!!!  #venezuelalibre  #venezuela  #sinluz  #feliznavidad  #feliznochebuena  #25diciembre  @freddyguevarac  @dolartoday    
-
-increible        llega  a  la  luz  a  las  6  45pm  (despues  de  estar  7horas  y  45min  #sinluz)    7  50pm  la  vuelven  a  quitar    a  q  juegan  sres  @corpoelecmerid_  @corpoelecinfo          #ejido  #meridasinluz    #singas  #singasolina  #sinagua
-
-otro  apagon    otro  dia  mas  con  el  socialismo  de  mierda  y  chavistas  vivos  #sinluz
-
-no  entremos  en  panico    desde  la  central  nuclear  de  atucha  dicen  que  esta  todo  bajo  control    fue  una  prueba  de  rutina  #sinluz    
-
-@claudiashein  @gobiernomx  importante  que  el  mundo  sepa    que  ustedes  seran  los  responsables  de  las  muertes  masivas  en  mexico@wsj  @guardian  @nytimes  @bbcmundo  @actualidadrt  @who  @onu_es  #coronavirus  #singobierno  #sinmedicinas    
-
-sin  luz    urgente    desd  las  8pm#parquecaiza  #16nov  colinasparquecaizacontrato  es100002181733  7poste53n57  cuenta10000159860contacto  04242588431  @corpoelecmiran_  @corpoelecinfo  @mppeevzla  #miranda  #venezuela  @alcaldejvra  @hectorodriguez  @fbritomaestre  #sinluz  @trafficmiranda
-
-nora  carpena  macri  vidal  romero  alberto  fernandez  #sinluzy  se  armo  nomas  en  las  fuerzas  armadas!designaron  un  bonafinista  en  su  obra  social  para  manotear  sus  $20  mil  millones  de  caja  y  ahogarlos  economicamente  y  hete  aqui  los  resultados  me  lo  acaba  de  mandar  un  alto  oficial      
-
-y  se  nos  fue  la  luz  #sinluz
-
-maracaibo  presenta  nuevo  apagon  que  afecta  a  decenas  de  familias  y  enfermos  en  los  ya  deterioradas  hospitales  la  culpa  no  es  de  la  vaca      es  del  chavismo  revolucion  destructora  #crisis  #zulia  #mcbo  #sinluz  #venezuela    
-
-a  dia  de  hoy  el  ministerio  de  seguridad  social  todavia  no  se  ha  puesto  en  contacto  con  las  mutuas  para  explicarles  como  han  de  tramitar  las  ayudas  a  los  autonomos    #apagon  #autonomos  #oingresos
-
-en  santa  cruz  de  mara  #sinluz  desde  las  9  pm    cdlm    @ldanieri
-
