--- brat-v1.3_Crunchy_Frog/data/second-iter/diego-sample_30-randstate_19-2020-06-15_202334.txt
@@ -1,60 +1,0 @@
-#sinluz  en  el  marques    alguien  mas  
-
-#23sep  |    nuestros  vecinos  en  #elhatillo  siguen  sufriendo  de  calamidad  del  servicio  electrico  sectores  como  #puebloelhatillo    #asoocho  #lasmarias    #oripoto    #launion    #losnaranjosrurales  y  #elencantado  estan  #sinluz  #corpoelecnosomosbuhos  @vpitv  @reporteya@monitorciudad_
-
-en  vela    otra  vez      @maienhaiku  #sinluz  en  5  7  5  #lunallena    #noviembre  #venezuela2019    
-
-#28nov  |  caf  dispone  de  $230  millones  para  enfrentar  crisis  electrica  en  4  estados  &gt;&gt;    
-
-#sinluz  #maracaibo  en  este  momento  sector  altos  de  jalisco  todos  los  circuitos  a  oscuras  @corpoeleczulia_    no  solucionan    
-
-cumpliendo  las  5  horas  #sinluz  en  los  #alacranes  #sanfelix      sin  informacion  @corpoelecsur  sectores  aledagnos  tienen  servicio  normal  @corpoelecinfo  restituyan  dicho  servicio  @corpoelec_boliv
-
-@prensacorpoelec  felices  que  desgraciados  el  sector  romulo    betancourt    en  #maturin  #sinluz    desde  el  dia  14  de  diciembre    aun  esperan  por  un  transformador    qu8en  puede  tener  feliz  navidad  asi  @yelitzepsuv_
-
-#sinluz  ya  casi  12  horas  sin  servicio    guama  yaracuy  
-
-es  sadismo  y  maldad  @corpoeleczulia_@fbritomaestre@nicolasmaduropues  si    otra  noche  #sinluz  para  media  #maracaibo  hasta  amanecer  #racionamientocriminal  el  calor  es  insoportable        lo  aplican  tambien  el  domingo  y  mas  o  menos  cuando  podemos  descansar  ah  #yabasta  #sos
-
-estoy  demasiado  molesta  e  indignada  con  esto  #sinluz    que  rabia    lo  unico  que  hacen  es  pisotear  al  venezolano  con  lo  que  le  de  la  gana    ruego  que  se  vayan  y  las  paguen  todas  
-
-usuarios  informan  que  estan  sin  electricidad    desde  la  1  pm  de  este  #19nov  zonas  de  los  estados    falcon    guarico    aragua    trujillo    zulia    miranda  y  yaracuyayudenos  a  informar  indicando  lugar    sitio  y  la  hora  #sinluz  
-
-@corpoelecinfo@corpoeleccar_sin  luz  urb  lomas  del  este  nic    1529725  telefono  02418584519    banco  mercantil  lomas  del  este  valencia  edo  carabobo  #sinluz2  50pm  23/10/19  2da  vez  que  se  va  la  luz@rafaellacava10  @traffivalencia
-
-@reporteya  #sinluz  en  maracaibo  como  de  esperar  
-
-#31jul  |  crisis  electrica  mantiene  de  rodillas  al  sector  comercial  larense  &gt;&gt;    
-
-ya  son  12hrs  mald      @corpoeleczulia_  @corpoelecinfo  #sinluz  en  el  sector  milagro  cotorrera  exigimos  el  mismo  trato  que  se  le  da  a  cerros  de  marin  sector  alimentado  por  la  misma  subestacion  @omarprietogob  @willy_casanova  #denuncia  @reporteya  @ldanieri  @gerardtorresp  @periodistaam
-
-@globalsonreir  calle  este  ii  de  la  morita  #sinluz  8  00am
-
-#ultimahora  #22j  |  se  registro  un  nuevo  #apagonnacional  los  estados  que  se  encuentran  #sinluz  son  guariconueva  espartala  guaira  distrito  capitalanzoateguiaraguamirandatachirameridacarabobozuliacojedesyaracuysucrefalconlaramonagas
-
-@oficialedesur  @pauliimp2  hoy  tres  cortes  en  la  misma  tarde  y  te  dicen  que  "siguen  por  privado"  para  no  decirte  nada    son  nefastos  #sinluz  #edesur
-
-#ahora  las  5  zonas  mas  afectadas  sin  suministro  electrico  por  empresa    #sinluz  #secorto  #josemariaezeiza  #barriolaunion    
-
-#19marla  tarde  de  este  jueves  se  registro  un  apagon  que  afecta  zonas  de  los  estados  #miranda    #vargas    #aragua  y  #caracas  #sinluz
-
-atencion!  se  viralizo  el  prende  y  apaga  de  #tn  #sinluz
-
-falso!  maraca  cuya    barrio  jose  felix  ribas    palo  verde  en  el  municipio  sucre  sin  luz  #sinluz#apagon    
-
-@hidrocaribe  @evelynb_vasquez  @mppaaguas  @hidrovenca  @maduro_en  @vicevenezuela  @nestorreverol  @luisjmarcano  @drodriguezven  lunes  04nov  hora  3  50pm    #sinagua;  los  tronconales    mcpo  bolivar    pqa  el  carmen    barcelona    anzoategui  @dcabellor  @luisjmarcano  @cle_anzoategui  @stellalugob  @fidelvasquepsuv  @nicolasmaduro
-
-tachira  esta  llevando  de  la  crisis  #apagon
-
-mi  amada  costa  oriental  del  lago  de  mi  juventud      #zulia  #cabimas  #sinluz  una  de  las  zonas  con  mayor  demanda  electrica  en  el  pais  y  muy  indefensa      ojala  te  pueda  implantar  una  gran  solucion          
-
-#sinluz  sectores  del  edo  #carabobo  se  encuentran  sin  energia  electrica
-
-quien  con  energia  electrica      sera  que  hay  un  nuevo  apagon      reporten  #sinluz  #cabudare
-
-@luisaguilarj  @corpoelecinfo  asi  es  luis    sin  luz  partes  altas  de  el  cafetal  y  caurimare  y  otras  zonas  cercanas  desde  hace  aproximadamente  media  hora    apreciamos  cualquier  informacion  luis    gracias  por  tu  atencion  y  tu  tiempo    #sinluz  @corpoelecinfo
-
-7  05  p  m      se  fue  la  luz    av    2  de  paraparal    diagonal  al  c  c    paseo  paraparal    los  guayos    carabobo    nic    1378691    tlf    0241  6150978    #sinluz  #carabobo  @corpoelecinfo  @corpoeleccar_  @rafaellacava10  @fbritomaestre  @reportelectrico  @alc_losguayos10  @lacava10dracula
-
-  @jguaido  pareces  un  vampiro    tienes  sed  de  sangre    quieres  lavar  la  sangre  en  las  calles  con  mas  sangre    montas  el  show  y  luego  huyes  y  los  pendejos  llevando  plomo  parejo  #el10pamiraflores  #5mar  #sinluz
-
