+++ c4v-py/reports/ml_models_performance_analysis/report-20200612_170023-balanced_dataset_brat.csv
@@ -1,0 +1,14 @@
+,accuracy,hamming_loss,f1_score_micro,f1_score_macro,precision_micro,precision_macro,recall_micro,recall_macro,model_name,model_class_name
+0,0.4166666666666667,0.15476190476190477,0.8,0.45568045399978174,0.7878787878787878,0.445578231292517,0.8125,0.5081632653061224,Multi kNN,MLkNN
+1,0.25,0.16666666666666666,0.787878787878788,0.3605805958747135,0.7647058823529411,0.3238095238095238,0.8125,0.40816326530612246,BR_Bayes BernoulliNB,BinaryRelevance
+2,0.25,0.16666666666666666,0.7941176470588235,0.36500341763499655,0.75,0.32142857142857145,0.84375,0.42857142857142855,BR_LogisticR,BinaryRelevance
+3,0.25,0.14285714285714285,0.8285714285714286,0.4792891319207109,0.7631578947368421,0.46428571428571436,0.90625,0.5238095238095238,BR_LinearSVC,BinaryRelevance
+4,0.25,0.16666666666666666,0.7941176470588235,0.36500341763499655,0.75,0.32142857142857145,0.84375,0.42857142857142855,BR_GradDesc SGDCClassifier,BinaryRelevance
+5,0.3333333333333333,0.19047619047619047,0.7142857142857143,0.2597402597402597,0.8333333333333334,0.2380952380952381,0.625,0.2857142857142857,LP_Bayes BernoulliNB,LabelPowerset
+6,0.3333333333333333,0.19047619047619047,0.7142857142857143,0.2597402597402597,0.8333333333333334,0.2380952380952381,0.625,0.2857142857142857,LP_LogisticR,LabelPowerset
+7,0.4166666666666667,0.15476190476190477,0.7796610169491525,0.4311232627022101,0.8518518518518519,0.48412698412698413,0.71875,0.41360544217687073,LP_LinearSVC,LabelPowerset
+8,0.5,0.13095238095238096,0.8196721311475409,0.4737076648841354,0.8620689655172413,0.4904761904761905,0.78125,0.48775510204081635,LP_GradDesc SGDCClassifier,LabelPowerset
+9,0.16666666666666666,0.17857142857142858,0.7692307692307692,0.3490259740259741,0.7575757575757576,0.3174603174603175,0.78125,0.38775510204081637,CC_Bayes BernoulliNB,ClassifierChain
+10,0.4166666666666667,0.13095238095238096,0.8196721311475409,0.3787878787878788,0.8620689655172413,0.380952380952381,0.78125,0.38775510204081637,CC_LogisticR,ClassifierChain
+11,0.3333333333333333,0.19047619047619047,0.7142857142857143,0.2597402597402597,0.8333333333333334,0.2380952380952381,0.625,0.2857142857142857,CC_LinearSVC,ClassifierChain
+12,0.4166666666666667,0.15476190476190477,0.7796610169491525,0.3454545454545454,0.8518518518518519,0.380952380952381,0.71875,0.3469387755102041,CC_GradDesc SGDCClassifier,ClassifierChain
