+++ c4v-py/reports/ml_models_performance_analysis/report-2020_06_26_200843-137t-sources.csv
@@ -1,0 +1,5 @@
+,source
+0,../brat-v1.3_Crunchy_Frog/data/first-iter/sampled_58_30
+1,../brat-v1.3_Crunchy_Frog/data/first-iter/balanced_dataset_brat
+2,../brat-v1.3_Crunchy_Frog/data/second-iter/diego-sample_30-randstate_19-2020-06-15_202334
+3,../brat-v1.3_Crunchy_Frog/data/second-iter/marianela-sample_50-randstate_42-2020-06-13_195818
