+++ c4v-py/src/c4v/__init__.py
@@ -1,0 +1,1 @@
+__version__ = '0.1.0'
