--- data_analysis/tagging-set-original_for_jupyter_tagging-annotated_ids.csv
@@ -1,101 +1,0 @@
-id
-1213216623114498048
-1153455759528779520
-1147221277729537920
-1153834864052822016
-1200281985815437056
-1140261255120400384
-1202301233161220096
-1217801564200546304
-1200226716825862144
-1140323498130366464
-1204762585238773504
-1243203000061091840
-1217052544867995648
-1230194202866221056
-1153649571773472768
-1234240956594966528
-1208401820533571584
-1171952925645709056
-1140333736296353792
-1233763856196894720
-1215617845054377984
-1187367872894509056
-1194380338602160128
-1140309222569271168
-1192620334354173952
-1244789353605455872
-1154847607098789888
-1213655266698440704
-1216329393448112128
-1153489054866038784
-1220516489461215232
-1168663646622441472
-1153625699258982400
-1166697330940284928
-1161376304542965504
-1173399103889432320
-1235705747364773888
-1140318872098279424
-1204503187014766592
-1153659540275003392
-1140224641568595968
-1164696994658574336
-1155541115728224256
-1235704041767923456
-1234233825682694144
-1230005154587516928
-1161145210975506432
-1241339673387192320
-1166782724511866880
-1151013866933018624
-1235689098527903744
-1214728364621205504
-1206174841704984320
-1188179844636512256
-1176592847761497856
-1153450825676709888
-1153467680105672704
-1156404098373160704
-1214342349939191808
-1214671525351186432
-1173743344662568704
-1209313560540729344
-1235725692018581504
-1198966585953865728
-1242153283155787520
-1175179597874417664
-1167492209375109120
-1202355587952066304
-1204222376608747520
-1140273626874359808
-1204893209282990080
-1197491187512487936
-1225213996254535680
-1199948088116428800
-1168390766366023680
-1202392591175692288
-1186459818443317248
-1162418697828540416
-1215945309123235840
-1175453803014959104
-1153426909885321216
-1196169610590138368
-1175145132632485888
-1219532632788070400
-1163914674972307456
-1185631580783546368
-1140348171572404224
-1178040109339426816
-1187575752956481536
-1229972654561660928
-1154132151836827648
-1204509185930616832
-1211704021616152320
-1226869695086813184
-1179938392290795520
-1175251669115949056
-1202300913026830336
-1232037590074826752
-1234813388334784256
-1182382288010665984
