--- helper_functions/ann_creator.sh
@@ -1,8 +1,0 @@
-#!/bin/bash
-
-for filename in *.txt
-do
-	filename=$(basename -- "$filename" .txt)
-	touch $filename.ann
-done
-
