--- helper_functions/reports/report-20200612_170024-sampled_58_30.csv
@@ -1,14 +1,0 @@
-,accuracy,hamming_loss,f1_score_micro,f1_score_macro,precision_micro,precision_macro,recall_micro,recall_macro,model_name,model_class_name
-0,0.0,0.25925925925925924,0.631578947368421,0.3246031746031746,0.5,0.2533068783068783,0.8571428571428571,0.5,Multi kNN,MLkNN
-1,0.0,0.3148148148148148,0.5853658536585367,0.2987012987012987,0.4444444444444444,0.22222222222222224,0.8571428571428571,0.5,BR_Bayes BernoulliNB,BinaryRelevance
-2,0.0,0.3148148148148148,0.5853658536585367,0.2987012987012987,0.4444444444444444,0.22222222222222224,0.8571428571428571,0.5,BR_LogisticR,BinaryRelevance
-3,0.0,0.3148148148148148,0.5853658536585367,0.2987012987012987,0.4444444444444444,0.22222222222222224,0.8571428571428571,0.5,BR_LinearSVC,BinaryRelevance
-4,0.0,0.3148148148148148,0.5853658536585367,0.2987012987012987,0.4444444444444444,0.22222222222222224,0.8571428571428571,0.5,BR_GradDesc SGDCClassifier,BinaryRelevance
-5,0.0,0.3148148148148148,0.5853658536585367,0.2987012987012987,0.4444444444444444,0.22222222222222224,0.8571428571428571,0.5,LP_Bayes BernoulliNB,LabelPowerset
-6,0.0,0.3148148148148148,0.5853658536585367,0.2987012987012987,0.4444444444444444,0.22222222222222224,0.8571428571428571,0.5,LP_LogisticR,LabelPowerset
-7,0.0,0.3148148148148148,0.5853658536585367,0.2987012987012987,0.4444444444444444,0.22222222222222224,0.8571428571428571,0.5,LP_LinearSVC,LabelPowerset
-8,0.0,0.3148148148148148,0.5853658536585367,0.2987012987012987,0.4444444444444444,0.22222222222222224,0.8571428571428571,0.5,LP_GradDesc SGDCClassifier,LabelPowerset
-9,0.0,0.3148148148148148,0.5853658536585367,0.2987012987012987,0.4444444444444444,0.22222222222222224,0.8571428571428571,0.5,CC_Bayes BernoulliNB,ClassifierChain
-10,0.0,0.3148148148148148,0.5853658536585367,0.2987012987012987,0.4444444444444444,0.22222222222222224,0.8571428571428571,0.5,CC_LogisticR,ClassifierChain
-11,0.0,0.3148148148148148,0.5853658536585367,0.2987012987012987,0.4444444444444444,0.22222222222222224,0.8571428571428571,0.5,CC_LinearSVC,ClassifierChain
-12,0.0,0.3148148148148148,0.5853658536585367,0.2987012987012987,0.4444444444444444,0.22222222222222224,0.8571428571428571,0.5,CC_GradDesc SGDCClassifier,ClassifierChain
