--- helper_functions/reports/report-2020_06_26_200843-137t-performance.csv
@@ -1,14 +1,0 @@
-,accuracy,hamming_loss,f1_score_micro,f1_score_macro,precision_micro,precision_macro,recall_micro,recall_macro,model_name,model_class_name
-9,0.11904761904761904,0.2108843537414966,0.7155963302752293,0.3359338986736247,0.7289719626168224,0.3046435965690624,0.7027027027027027,0.3766233766233766,CC_Bayes BernoulliNB,ClassifierChain
-0,0.14285714285714285,0.19727891156462585,0.7264150943396227,0.5673080630662598,0.7623762376237624,0.6260049474335189,0.6936936936936937,0.5281983761246434,Multi kNN,MLkNN
-1,0.14285714285714285,0.20748299319727892,0.7239819004524887,0.34228310502283105,0.7272727272727273,0.30559916274201987,0.7207207207207207,0.38961038961038963,BR_Bayes BernoulliNB,BinaryRelevance
-10,0.16666666666666666,0.23129251700680273,0.653061224489796,0.24704500978473584,0.7529411764705882,0.21768707482993196,0.5765765765765766,0.2857142857142857,CC_LogisticR,ClassifierChain
-2,0.19047619047619047,0.21768707482993196,0.728813559322034,0.3468182524151214,0.688,0.29434212709474034,0.7747747747747747,0.42857142857142855,BR_LogisticR,BinaryRelevance
-4,0.19047619047619047,0.20408163265306123,0.7413793103448275,0.35349206349206347,0.7107438016528925,0.3029656021587069,0.7747747747747747,0.42857142857142855,BR_GradDesc SGDCClassifier,BinaryRelevance
-5,0.19047619047619047,0.22789115646258504,0.6564102564102564,0.24704500978473584,0.7619047619047619,0.21768707482993196,0.5765765765765766,0.2857142857142857,LP_Bayes BernoulliNB,LabelPowerset
-6,0.19047619047619047,0.22448979591836735,0.6597938144329896,0.24873015873015872,0.7710843373493976,0.22025883524141365,0.5765765765765766,0.2857142857142857,LP_LogisticR,LabelPowerset
-3,0.21428571428571427,0.20068027210884354,0.7445887445887446,0.3552676889965026,0.7166666666666667,0.3052009201834986,0.7747747747747747,0.42857142857142855,BR_LinearSVC,BinaryRelevance
-7,0.21428571428571427,0.22108843537414966,0.6829268292682927,0.35370599852844364,0.7446808510638298,0.42364275291104564,0.6306306306306306,0.35262832405689554,LP_LinearSVC,LabelPowerset
-8,0.21428571428571427,0.20748299319727892,0.7081339712918658,0.42085417800862235,0.7551020408163265,0.4998556998556999,0.6666666666666666,0.3991601332615158,LP_GradDesc SGDCClassifier,LabelPowerset
-11,0.21428571428571427,0.2108843537414966,0.6990291262135923,0.31595704948646125,0.7578947368421053,0.3154969304795089,0.6486486486486487,0.33766233766233766,CC_LinearSVC,ClassifierChain
-12,0.21428571428571427,0.21768707482993196,0.6767676767676768,0.2816971916971917,0.7701149425287356,0.3274016923842708,0.6036036036036037,0.3051948051948052,CC_GradDesc SGDCClassifier,ClassifierChain
