--- helper_functions/reports/report-2020_06_28_072853-148t-performance.csv
@@ -1,14 +1,0 @@
-,accuracy,hamming_loss,f1_score_micro,f1_score_macro,precision_micro,precision_macro,recall_micro,recall_macro,model_name,model_class_name
-1,0.15555555555555556,0.2,0.7449392712550608,0.34551673919588727,0.6814814814814815,0.292063492063492,0.8214285714285714,0.42857142857142855,BR_Bayes BernoulliNB,BinaryRelevance
-2,0.15555555555555556,0.19365079365079366,0.7510204081632653,0.3485494280828321,0.6917293233082706,0.2959025470653378,0.8214285714285714,0.42857142857142855,BR_LogisticR,BinaryRelevance
-5,0.15555555555555556,0.2222222222222222,0.6534653465346534,0.2408889725962897,0.7333333333333333,0.20952380952380953,0.5892857142857143,0.2857142857142857,LP_Bayes BernoulliNB,LabelPowerset
-6,0.15555555555555556,0.2222222222222222,0.6534653465346534,0.2408889725962897,0.7333333333333333,0.20952380952380953,0.5892857142857143,0.2857142857142857,LP_LogisticR,LabelPowerset
-9,0.15555555555555556,0.19682539682539682,0.7479674796747967,0.34701142157588155,0.6865671641791045,0.29393939393939394,0.8214285714285714,0.42857142857142855,CC_Bayes BernoulliNB,ClassifierChain
-10,0.15555555555555556,0.19365079365079366,0.7510204081632653,0.3485494280828321,0.6917293233082706,0.2959025470653378,0.8214285714285714,0.42857142857142855,CC_LogisticR,ClassifierChain
-3,0.17777777777777778,0.1873015873015873,0.7591836734693878,0.3738927753439876,0.6992481203007519,0.4408520420148327,0.8303571428571429,0.44155844155844154,BR_LinearSVC,BinaryRelevance
-4,0.2,0.18412698412698414,0.7622950819672133,0.37546920500260905,0.7045454545454546,0.4430417128091547,0.8303571428571429,0.44155844155844154,BR_GradDesc SGDCClassifier,BinaryRelevance
-11,0.2,0.17142857142857143,0.773109243697479,0.3814968367613063,0.7301587301587301,0.38508554937126366,0.8214285714285714,0.43606393606393606,CC_LinearSVC,ClassifierChain
-0,0.24444444444444444,0.18412698412698414,0.7521367521367521,0.5374408504660606,0.7213114754098361,0.5658163265306123,0.7857142857142857,0.5697453525039732,Multi kNN,MLkNN
-8,0.24444444444444444,0.19682539682539682,0.7019230769230769,0.3320411091265774,0.7604166666666666,0.4031746031746032,0.6517857142857143,0.33337352302869544,LP_GradDesc SGDCClassifier,LabelPowerset
-12,0.26666666666666666,0.15555555555555556,0.7896995708154506,0.3912905308925848,0.7603305785123967,0.4004417771109444,0.8214285714285714,0.43606393606393606,CC_GradDesc SGDCClassifier,ClassifierChain
-7,0.28888888888888886,0.1746031746031746,0.7393364928909952,0.41259763164525076,0.7878787878787878,0.5986432266920072,0.6964285714285714,0.390278686830411,LP_LinearSVC,LabelPowerset
