+++ old_src/c4v_py.egg-info/SOURCES.txt
@@ -1,0 +1,11 @@
+README.md
+setup.py
+src/c4v/__init__.py
+src/c4v/data/baseline_models.py
+src/c4v/data/data_loader.py
+src/c4v/data/data_sampler.py
+src/c4v_py.egg-info/PKG-INFO
+src/c4v_py.egg-info/SOURCES.txt
+src/c4v_py.egg-info/dependency_links.txt
+src/c4v_py.egg-info/requires.txt
+src/c4v_py.egg-info/top_level.txt
