+++ old_src/c4v_py.egg-info/requires.txt
@@ -1,0 +1,9 @@
+jupyter<2.0.0,>=1.0.0
+nltk<4.0,>=3.5
+pandas==1.0.5
+pip<21.0.0,>=20.0.0
+scikit-learn==0.23.1
+scikit-multilearn==0.2.0
+tensorflow-probability==0.10.0
+tensorflow==2.2.0
+tensorflow_hub[make_image_classifier]==0.8.0
