--- tests/test_c4v_py.py
@@ -1,5 +1,0 @@
-from c4v import __version__
-
-
-def test_version():
-    assert __version__ == '0.1.0'
